HARP-12587: Moves TileObject rendering to TileObjectRenderer class
