+++ @here/harp-mapview/lib/MapView.ts
@@ -5,12 +5,8 @@
  */
 import {
     Env,
-    Expr,
-    getFeatureId,
-    getPropertyValue,
     GradientSky,
     ImageTexture,
-    IndexedTechnique,
     Light,
     MapEnv,
     PostEffects,
@@ -61,12 +57,10 @@ import { ElevationRangeSource } from "./ElevationRangeSource";
 import { EventDispatcher } from "./EventDispatcher";
 import { FrustumIntersection } from "./FrustumIntersection";
 import { overlayOnElevation } from "./geometry/overlayOnElevation";
-import { SolidLineMesh } from "./geometry/SolidLineMesh";
 import { TileGeometryManager } from "./geometry/TileGeometryManager";
 import { MapViewImageCache } from "./image/MapViewImageCache";
 import { IntersectParams } from "./IntersectParams";
 import { MapAnchors } from "./MapAnchors";
-import { MapObjectAdapter } from "./MapObjectAdapter";
 import { MapViewFog } from "./MapViewFog";
 import { MapViewTaskScheduler } from "./MapViewTaskScheduler";
 import { PickHandler, PickResult } from "./PickHandler";
@@ -87,7 +81,8 @@ import { TextElementsRenderer, ViewUpdateCallback } from "./text/TextElementsRen
 import { TextElementsRendererOptions } from "./text/TextElementsRendererOptions";
 import { createLight } from "./ThemeHelpers";
 import { ThemeLoader } from "./ThemeLoader";
-import { Tile, TileFeatureData, TileObject } from "./Tile";
+import { Tile } from "./Tile";
+import { TileObjectRenderer } from "./TileObjectsRenderer";
 import { MapViewUtils } from "./Utils";
 import { ResourceComputationType, VisibleTileSet, VisibleTileSetOptions } from "./VisibleTileSet";
 
@@ -157,13 +152,6 @@ const DEFAULT_CAM_FAR_PLANE = 4000000;
 const MAX_FIELD_OF_VIEW = 140;
 const MIN_FIELD_OF_VIEW = 10;
 
-/**
- * All objects in fallback tiles are reduced by this amount.
- *
- * @internal
- */
-export const FALLBACK_RENDER_ORDER_OFFSET = 20000;
-
 const DEFAULT_MIN_ZOOM_LEVEL = 1;
 
 /**
@@ -180,8 +168,6 @@ const DEFAULT_MIN_CAMERA_HEIGHT = 20;
  * Style set used by {@link PolarTileDataSource} by default.
  */
 const DEFAULT_POLAR_STYLE_SET_NAME = "polar";
-
-const DEFAULT_STENCIL_VALUE = 1;
 
 /**
  * The type of `RenderEvent`.
@@ -623,9 +609,10 @@ export interface MapViewOptions extends TextElementsRendererOptions, Partial<Loo
     enableMixedLod?: boolean;
 
     /**
-     * If enableMixedLod is true, this value will be used to calculate the minimum Pixel Size of a
-     * Tile regarding to the screen size, when the subdivision of tiles is stopped and therefore
-     * higher level tiles rendered.
+     * If enableMixedLod is `true`, this value will be used to calculate the minimum Pixel Size of a
+     * tile regarding to the screen size. When the area of a tile is smaller then this calculated
+     * area on the screen, the subdivision of tiles is stopped and therefore higher level tiles will
+     * be rendered instead.
      * @beta
      *
      * @default 256
@@ -829,6 +816,7 @@ export class MapView extends EventDispatcher {
         | ScreenCollisionsDebug = new ScreenCollisions();
 
     private m_visibleTiles: VisibleTileSet;
+    private readonly m_tileObjectRenderer: TileObjectRenderer;
 
     private m_elevationSource?: DataSource;
     private m_elevationRangeSource?: ElevationRangeSource;
@@ -936,6 +924,7 @@ export class MapView extends EventDispatcher {
 
     private readonly m_imageCache: MapViewImageCache = new MapViewImageCache(this);
     private readonly m_userImageCache: MapViewImageCache = new MapViewImageCache(this);
+    private readonly m_env: MapEnv = new MapEnv({});
 
     private readonly m_poiManager: PoiManager = new PoiManager(this);
 
@@ -953,14 +942,9 @@ export class MapView extends EventDispatcher {
     private m_copyrightInfo: CopyrightInfo[] = [];
     private readonly m_animatedExtrusionHandler: AnimatedExtrusionHandler;
 
-    private readonly m_env: MapEnv = new MapEnv({});
-
     private m_enableMixedLod: boolean | undefined;
     private readonly m_lodMinTilePixelSize: number | undefined;
 
-    private readonly m_renderOrderStencilValues = new Map<number, number>();
-    // Valid values start at 1, because the screen is cleared to zero
-    private m_stencilValue: number = DEFAULT_STENCIL_VALUE;
     private m_taskScheduler: MapViewTaskScheduler;
 
     // `true` if dispose() has been called on `MapView`.
@@ -1133,6 +1117,7 @@ export class MapView extends EventDispatcher {
         // setup camera with initial position
         this.setupCamera();
 
+        this.m_tileObjectRenderer = new TileObjectRenderer(this.m_env);
         this.m_raycaster = new PickingRaycaster(width, height, this.m_env);
 
         this.m_movementDetector = new CameraMovementDetector(
@@ -3040,27 +3025,6 @@ export class MapView extends EventDispatcher {
         return this.m_fog;
     }
 
-    private getStencilValue(renderOrder: number) {
-        if (!this.m_drawing) {
-            throw new Error("failed to get the stencil value");
-        }
-
-        return (
-            this.m_renderOrderStencilValues.get(renderOrder) ??
-            this.allocateStencilValue(renderOrder)
-        );
-    }
-
-    private allocateStencilValue(renderOrder: number) {
-        if (!this.m_drawing) {
-            throw new Error("failed to allocate stencil value");
-        }
-
-        const stencilValue = this.m_stencilValue++;
-        this.m_renderOrderStencilValues.set(renderOrder, stencilValue);
-        return stencilValue;
-    }
-
     private setPostEffects() {
         // First clear all the effects, then enable them from what is specified.
         this.mapRenderingManager.bloom.enabled = false;
@@ -3615,8 +3579,7 @@ export class MapView extends EventDispatcher {
         this.RENDER_EVENT.time = frameStartTime;
         this.dispatchEvent(this.RENDER_EVENT);
 
-        this.m_stencilValue = DEFAULT_STENCIL_VALUE;
-        this.m_renderOrderStencilValues.clear();
+        this.m_tileObjectRenderer.prepareRender();
 
         ++this.m_frameNumber;
 
@@ -3701,7 +3664,12 @@ export class MapView extends EventDispatcher {
         // no need to check everything if we're not going to create text renderer.
         renderList.forEach(({ zoomLevel, renderedTiles }) => {
             renderedTiles.forEach(tile => {
-                this.renderTileObjects(tile, zoomLevel);
+                this.m_tileObjectRenderer.render(
+                    tile,
+                    zoomLevel,
+                    this.m_camera.position,
+                    this.m_sceneRoot
+                );
 
                 //We know that rendered tiles are visible (in the view frustum), so we update the
                 //frame number, note we don't do this for the visibleTiles because some may still be
@@ -3888,219 +3856,6 @@ export class MapView extends EventDispatcher {
             this.FRAME_COMPLETE_EVENT.time = frameStartTime;
             this.dispatchEvent(this.FRAME_COMPLETE_EVENT);
         }
-    }
-
-    private renderTileObjects(tile: Tile, zoomLevel: number) {
-        const worldOffsetX = tile.computeWorldOffsetX();
-        if (tile.willRender(zoomLevel)) {
-            for (const object of tile.objects) {
-                const mapObjectAdapter = MapObjectAdapter.get(object);
-                if (!this.processTileObject(tile, object, mapObjectAdapter)) {
-                    continue;
-                }
-
-                this.updateStencilRef(object);
-
-                object.position.copy(tile.center);
-                if (object.displacement !== undefined) {
-                    object.position.add(object.displacement);
-                }
-                object.position.x += worldOffsetX;
-                object.position.sub(this.m_camera.position);
-                if (tile.localTangentSpace) {
-                    object.setRotationFromMatrix(tile.boundingBox.getRotationMatrix());
-                }
-                object.frustumCulled = false;
-
-                this.adjustRenderOrderForFallback(object, mapObjectAdapter, tile);
-                this.m_sceneRoot.add(object);
-            }
-            tile.didRender();
-        }
-    }
-
-    private updateStencilRef(object: TileObject) {
-        // TODO: acquire a new style value of if transparent
-        if (object.renderOrder !== undefined && object instanceof SolidLineMesh) {
-            const material = object.material;
-            if (Array.isArray(material)) {
-                material.forEach(
-                    mat => (mat.stencilRef = this.getStencilValue(object.renderOrder))
-                );
-            } else {
-                material.stencilRef = this.getStencilValue(object.renderOrder);
-            }
-        }
-    }
-
-    private adjustRenderOrderForFallback(
-        object: TileObject,
-        mapObjectAdapter: MapObjectAdapter | undefined,
-        tile: Tile
-    ) {
-        // When falling back to a parent tile (i.e. tile.levelOffset < 0) there will
-        // be overlaps with the already loaded tiles. Therefore all (flat) objects
-        // in a fallback tile must be shifted, such that their renderOrder is less
-        // than the groundPlane that each neighbouring Tile has (it has a renderOrder
-        // of -10000, see addGroundPlane in TileGeometryCreator), only then can we be
-        // sure that nothing of the parent will be rendered on top of the children,
-        // as such, we shift using the FALLBACK_RENDER_ORDER_OFFSET.
-        // This does not apply to buildings b/c they are 3d and the overlaps
-        // are resolved with a depth prepass. Note we set this always to ensure that if
-        // the Tile is used as a fallback, and then used normally, that we have the correct
-        // renderOrder.
-
-        if (tile.levelOffset >= 0) {
-            if (object._backupRenderOrder !== undefined) {
-                // We messed up the render order when this tile was used as fallback.
-                // Now we render normally, so restore the original renderOrder.
-                object.renderOrder = object._backupRenderOrder;
-            }
-            return;
-        }
-        let offset = FALLBACK_RENDER_ORDER_OFFSET;
-        const technique = mapObjectAdapter?.technique;
-        if (technique?.name === "extruded-polygon") {
-            // Don't adjust render order for extruded-polygon b/c it's not flat.
-            return;
-        }
-
-        if ((technique as any)?._category?.startsWith("road") === true) {
-            // Don't adjust render order for roads b/c the outline of the child tile
-            // would overlap the outline of the fallback parent.
-            // Road geometry would be duplicated but since it's rendered with two passes
-            // it would just appear a bit wider. That artefact is not as disturbing
-            // as seeing the cap outlines.
-            // NOTE: Since our tests do pixel perfect image comparison we also need to add a
-            // tiny offset in this case so that the order is well defined.
-            offset = 1e-6;
-        }
-
-        if (object._backupRenderOrder === undefined) {
-            object._backupRenderOrder = object.renderOrder;
-        }
-        object.renderOrder = object._backupRenderOrder + offset * tile.levelOffset;
-    }
-
-    /**
-     * Process dynamic updates of [[TileObject]]'s style.
-     *
-     * @returns `true` if object shall be used in scene, `false` otherwise
-     */
-    private processTileObject(tile: Tile, object: TileObject, mapObjectAdapter?: MapObjectAdapter) {
-        if (!object.visible) {
-            return false;
-        }
-        if (!this.processTileObjectFeatures(tile, object)) {
-            return false;
-        }
-
-        if (mapObjectAdapter) {
-            mapObjectAdapter.ensureUpdated(this);
-            if (!mapObjectAdapter.isVisible()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Process the features owned by the given [[TileObject]].
-     *
-     * @param tile - The {@link Tile} owning the [[TileObject]]'s features.
-     * @param object - The [[TileObject]] to process.
-     * @returns `false` if the given [[TileObject]] should not be added to the scene.
-     */
-    private processTileObjectFeatures(tile: Tile, object: TileObject): boolean {
-        const technique: IndexedTechnique = object.userData.technique;
-
-        if (!technique || technique.enabled === undefined) {
-            // Nothing to do, there's no technique.
-            return true;
-        }
-
-        const feature: TileFeatureData = object.userData.feature;
-
-        if (!feature || !Expr.isExpr(technique.enabled)) {
-            return Boolean(getPropertyValue(technique.enabled, this.m_env));
-        }
-
-        const { starts, objInfos } = feature;
-
-        if (!Array.isArray(objInfos) || !Array.isArray(starts)) {
-            // Nothing to do, the object is missing feature ids and their position
-            // in the index buffer.
-            return true;
-        }
-
-        const geometry: THREE.BufferGeometry | undefined = (object as any).geometry;
-
-        if (!geometry || !geometry.isBufferGeometry) {
-            // Nothing to do, the geometry is not a [[THREE.BufferGeometry]]
-            // and we can't generate groups.
-            return true;
-        }
-
-        // ExtrudeBufferGeometry for example doesn't have an index, hence we get the final index
-        // from the number of vertices.
-        const finalIndex = geometry.getIndex()?.count ?? geometry.attributes.position.count;
-
-        // clear the groups.
-        geometry.clearGroups();
-
-        // The offset in the index buffer of the end of the last
-        // pushed group.
-        let endOfLastGroup: number | undefined;
-
-        objInfos.forEach((properties, featureIndex) => {
-            // the id of the current feature.
-            const featureId = getFeatureId(properties);
-
-            let enabled = true;
-
-            if (Expr.isExpr(technique.enabled)) {
-                // the state of current feature.
-                const featureState = tile.dataSource.getFeatureState(featureId);
-
-                // create a new {@link @here/harp-datasource-protocol#Env} that can be used
-                // to evaluate expressions that access the feature state.
-                const $state = featureState ? new MapEnv(featureState) : null;
-
-                const parentEnv =
-                    typeof properties === "object"
-                        ? new MapEnv(properties, this.m_env)
-                        : this.m_env;
-
-                const env = new MapEnv({ $state }, parentEnv);
-
-                enabled = Boolean(getPropertyValue(technique.enabled, env));
-            }
-
-            if (!enabled) {
-                // skip this feature, it was disabled.
-                return;
-            }
-
-            // HARP-12247, geometry with no featureStarts would set start to `undefined`, in this
-            // case, `endOfLastGroup` is also undefined (first execution in this loop), so it would
-            // try to change the count of a group which hasn't yet been added, `addGroup` wasn't yet
-            // called, hence we use the `??` operator and fall back to 0. Because featureStarts are
-            // optional, we need to have a fallback.
-            const start = starts[featureIndex] ?? 0;
-            const end = starts[featureIndex + 1] ?? finalIndex;
-            const count = end - start;
-
-            if (start === endOfLastGroup) {
-                // extend the last group
-                geometry.groups[geometry.groups.length - 1].count += count;
-            } else {
-                geometry.addGroup(start, count);
-            }
-
-            endOfLastGroup = start + count;
-        });
-
-        return geometry.groups.length > 0;
     }
 
     private prepareRenderTextElements(time: number) {
